+++ src/common/CreatioClient/enums.ts
@@ -119,3 +119,14 @@ export enum ColumnUsage{
 	 */
 	None = 2
 }
+
+
+export enum LogLevel{
+	ALL,
+	Debug,
+	Error,
+	Fatal,
+	Info,
+	Trace,
+	Warn
+}